+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
 ### [0.5.7](https://github.com/mangata-finance/mangata-SDK/compare/v0.5.6...v0.5.7) (2021-10-12)
+- fix issue with decimals: from string to number
 
 ### [0.5.6](https://github.com/mangata-finance/mangata-SDK/compare/v0.5.5...v0.5.6) (2021-10-12)
 - dry bump